Bug3 Fix
Random value to compare against drop rates was created new in every iteration of the SelectDropRandomly For-loop. Probabilities of 0.5 are constantly checked against a newly created random value of 0-1, which leads to items not being dropped at all, or the first iteration leading to a drop.
To fix this, I moved the random number generation outside of the for loop and subtratcted the item[i] probability in each iteration.